+++ src/Logger.ts
@@ -2,7 +2,7 @@ import randomColor from 'randomcolor';
 import Style from './types/Style';
 import ComponentLogger from './ComponentLogger';
 import { Options } from './types/Options';
-import BrowserLogger from './BrowserLogger';
+import BrowserLogger from './loggers/BrowserLogger';
 
 const componentBackgroundColor = 'lightgray';